updating examples to include demonstration of inference

+++ t/unit/read/test_read.c
@@ -26,7 +26,7 @@ end\n\
 \n\
 # simple function to test return values\n\
 fn add_one(i::Int) -> Int\n\
-    let tmp::Int = i\n\
+    let tmp = i\n\
     tmp = i + 1\n\
     return tmp\n\
 end\n\
